pyproject.toml
[project]
name = "checkio"
version = "0.1.0"
description = "Checkio - A tool that not only tests your code but also suggests improvements and error fixes, making lab evaluations smarter and faster."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "build>=1.2.2.post1",
    "click>=8.1.8",
    "flask-cors>=5.0.0",
    "flask>=3.1.0",
    "pyjwt>=2.10.1",
    "rapidfuzz>=3.12.1",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.38",
    "textual>=1.0.0",
]
